Backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==64.0.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
optional-django==0.1.0
PyJWT==2.9.0
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
wheel==0.44.0